update mobile layout of transfer table
