+++ src/pages/Bridge/BridgeTransferHistory/TimeStatusCell.tsx
@@ -3,8 +3,7 @@ import dayjs from 'dayjs'
 import { Flex, Text } from 'rebass'
 
 import useTheme from 'hooks/useTheme'
-
-import { FULL_DATE_FORMAT } from '../consts'
+import { FULL_DATE_FORMAT } from 'pages/Bridge/consts'
 
 type Props = {
   timestamp?: number | ''
